Create and use new component CustomListHeader
