+++ src/components/SelectionListWithModal/CustomListHeader.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import {View} from 'react-native';
+import Text from '@components/Text';
+import useStyleUtils from '@hooks/useStyleUtils';
+import useThemeStyles from '@hooks/useThemeStyles';
+
+type CustomListHeaderProps = {
+    canSelectMultiple: Boolean | undefined;
+    leftHeaderText?: String | undefined;
+    rightHeaderText?: String | undefined;
+};
+
+function CustomListHeader({canSelectMultiple, leftHeaderText = '', rightHeaderText = ''}: CustomListHeaderProps) {
+    const styles = useThemeStyles();
+    const StyleUtils = useStyleUtils();
+
+    const header = (
+        <View
+            style={[
+                styles.flex1,
+                styles.flexRow,
+                styles.justifyContentBetween,
+                // Required padding accounting for the checkbox in multi-select mode
+                canSelectMultiple && styles.pl3,
+            ]}
+        >
+            <Text style={styles.searchInputStyle}>{leftHeaderText}</Text>
+            <View style={[StyleUtils.getMinimumWidth(60)]}>
+                <Text style={[styles.searchInputStyle, styles.textAlignCenter]}>{rightHeaderText}</Text>
+            </View>
+        </View>
+    );
+    
+    if (canSelectMultiple) {
+        return header;
+    }
+    return <View style={[styles.flexRow, styles.ph9, styles.pv3, styles.pb5]}>{header}</View>;
+}
+
+export default CustomListHeader;
